实现了沉浸式状态栏，部分activity改为了toolbar，侧边抽屉栏也都基本做好了，接下来要写BaseActivity

+++ MPChartLib/src/com/github/mikephil/charting/charts/RadarChart.java
@@ -216,7 +216,7 @@ public class RadarChart extends PieRadarChartBase<RadarData> {
      */
     public float getFactor() {
         RectF content = mViewPortHandler.getContentRect();
-        return (float) Math.min(content.width() / 2f, content.height() / 2f)
+        return Math.min(content.width() / 2f, content.height() / 2f)
                 / mYAxis.mAxisRange;
     }
