+++ MPChartLib/src/com/github/mikephil/charting/utils/Transformer.java
@@ -46,8 +46,8 @@ public class Transformer {
      */
     public void prepareMatrixValuePx(float xChartMin, float deltaX, float deltaY, float yChartMin) {
 
-        float scaleX = (float) ((mViewPortHandler.contentWidth()) / deltaX);
-        float scaleY = (float) ((mViewPortHandler.contentHeight()) / deltaY);
+        float scaleX = (mViewPortHandler.contentWidth()) / deltaX;
+        float scaleY = (mViewPortHandler.contentHeight()) / deltaY;
 
         // setup all matrices
         mMatrixValueToPx.reset();
