+++ MPChartLib/src/com/github/mikephil/charting/utils/Utils.java
@@ -387,7 +387,7 @@ public abstract class Utils {
 
             if (axis == null || sel.dataSet.getAxisDependency() == axis) {
 
-                float cdistance = Math.abs((float) sel.val - val);
+                float cdistance = Math.abs(sel.val - val);
                 if (cdistance < distance) {
                     index = valsAtIndex.get(i).dataSetIndex;
                     distance = cdistance;
